+++ CHANGELOG.md
@@ -73,11 +73,13 @@
 
 ## [Unreleased][]
 ### Added
+- UI: new 'Lines (high quality)' chart type. See [here](https://hydrogenaudio.org/index.php/topic,126743.msg1086993.html#msg1086993), uses DrawLines + PushClip method.
 ### Changed
 - Helpers: updated helpers.
 ### Removed
 ### Fixed
 - UI: chart types not being changed when using menu settings on statistics mode. See [here](https://hydrogenaudio.org/index.php/topic,125040.msg1086767.html#msg1086767).
+- UI: minor color errors on gradient charts when values were zero in some cases.
 - Drag n' drop: fix crash in some cases while scrolling to the end of the list while drag n' dropping.
 - Fixed UTF-8 script files without BOM not being properly recognized on windows.